TCVP-2397 Add ability to use basic auth to obtain token instead of openid connect
